EmailService refactoring. This close #773
