+++ restcomm/restcomm.email.api/src/main/java/org/mobicents/servlet/restcomm/api/EmailResponse.java
@@ -17,21 +17,24 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>
  *
  */
-package org.mobicents.servlet.restcomm.email;
+package org.mobicents.servlet.restcomm.api;
+
+
+import org.mobicents.servlet.restcomm.patterns.StandardResponse;
 
 /**
  * @author liblefty@gmail.com (Lefteris Banos)
  */
-public class EmailRequest {
-    private final Mail emailmsg;
-
-    public  EmailRequest(Mail object){
-        super();
-        this.emailmsg=object;
+public class EmailResponse<T> extends StandardResponse <T> {
+    public EmailResponse(final T object) {
+        super(object);
     }
 
-    public Mail getObject(){
-        return this.emailmsg;
+    public EmailResponse(final Throwable cause) {
+        super(cause);
     }
 
+    public EmailResponse(final Throwable cause, final String message) {
+        super(cause, message);
+    }
 }
