+++ restcomm/restcomm.email.api/src/main/java/org/mobicents/servlet/restcomm/api/Mail.java
@@ -17,7 +17,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>
  *
  */
-package org.mobicents.servlet.restcomm.email;
+package org.mobicents.servlet.restcomm.api;
 
 import org.joda.time.DateTime;
 import org.mobicents.servlet.restcomm.annotations.concurrency.Immutable;
