+++ restcomm/restcomm.email/pom.xml
@@ -58,5 +58,24 @@
                 <scope>provided</scope>
             </dependency>
 
+            <dependency>
+                <groupId>com.icegreen</groupId>
+                <artifactId>greenmail</artifactId>
+                <version>1.4.0</version>
+                <scope>test</scope>
+            </dependency>
+
+            <dependency>
+                <groupId>com.telestax.servlet</groupId>
+                <artifactId>restcomm.email.api</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+
+            <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-log4j12</artifactId>
+            <version>1.5.5</version>
+            </dependency>
+
         </dependencies>
     </project>
