+++ restcomm/restcomm.email/src/test/resources/emailServiceTest.xml
